backend/requirements.txt
asgiref
Django~=5.1.5
django-cors-headers
djangorestframework~=3.15.2
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv~=1.0.1
mercadopago~=2.2.3
requests~=2.32.3
django-allauth~=65.4.0
google~=3.0.0